requirements.txt
# Core Data Science Libraries
pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.0
xgboost==2.0.0

# API and Web Framework
google-api-python-client==2.100.0
flask==3.0.0
flask-cors==4.0.0

# Data Processing
python-dotenv==1.0.0
joblib==1.3.2

# Visualization (optional, for notebooks)
matplotlib==3.9.0
seaborn==0.12.2
plotly==5.17.0

# Utilities
requests==2.31.0
tqdm==4.66.1


